Make startup lazier for unit tests.
Wasn't able to test elsewhere stuff without this commit since it was
trying to load network data before the database was setup in a unit test
environment. Made it so it would lazily retry if it could not get data.

Also made the network data be stored and accessed as dicts instead of
lists to avoid linear searches.

Removed some print statements.